test proc file access

+++ ligerosmart-easyinstall.sh
@@ -101,15 +101,17 @@ if [ $install_dockercompose ]; then
 fi;
 
 ## Kernel config for elasticsearch
-max_map_count=$(sysctl -n vm.max_map_count || echo 0)
-if [ $max_map_count -lt 262144 ]; then
-    echo "The elasticsearch service will not run with max_map_count=$max_map_count. I will try to increase it"
-    sysctl -w vm.max_map_count=262144
-    echo 'vm.max_map_count=262144' > /etc/sysctl.d/elasticsearch.conf
-fi
+if [ -f /proc/sys/vm/max_map_count ]; then
+    max_map_count=$(sysctl -n vm.max_map_count || echo 0)
+    if [ $max_map_count -lt 262144 ]; then
+        echo "The elasticsearch service will not run with max_map_count=$max_map_count. I will try to increase it"
+        sysctl -w vm.max_map_count=262144
+        echo 'vm.max_map_count=262144' > /etc/sysctl.d/elasticsearch.conf
+    fi
 
-if [ $installation_only ]; then 
-    exit 0; 
+    if [ $installation_only ]; then 
+        exit 0; 
+    fi;
 fi;
 
 # Stack repository
